F4 : USB CDC : small correction for previous commit

+++ STM32F4/cores/maple/libmaple/usbF4/usb.c
@@ -36,7 +36,7 @@ void setupUSB (void)
 
 extern uint16_t VCP_DataTx (const uint8_t* Buf, uint32_t Len);
 extern uint8_t  VCPGetByte(void);
-extern uint8_t  VCPGetBytes(uint8_t * rxBuf, uint32_t len);
+extern uint32_t VCPGetBytes(uint8_t * rxBuf, uint32_t len);
 
 uint32_t usbSendBytes(const uint8_t* sendBuf, uint32_t len)
 {
